integration complete

+++ pixsfm/base/src/projection.h
@@ -102,9 +102,9 @@ inline void WorldToPixel(const colmap::Camera& camera,
                          const Eigen::Vector4d& qvec,
                          const Eigen::Vector3d& tvec,
                          const Eigen::Vector3d& xyz, double* xy) {
-  switch (camera.model_id) {
+  switch (static_cast<int>(camera.model_id)) {
 #define CAMERA_MODEL_CASE(CameraModel)                                  \
-  case colmap::CameraModel::model_id:                                   \
+  case static_cast<int>(colmap::CameraModel::model_id):                                   \
     WorldToPixel<colmap::CameraModel>(&camera.params[0], qvec.data(), \
                                       tvec.data(), xyz.data(), xy);     \
     break;
